Move RateLimitContainer to internal package
