+++ src/main/kotlin/io/github/freya022/botcommands/internal/commands/ratelimit/declaration/RateLimitManagerImpl.kt
@@ -1,13 +1,13 @@
 package io.github.freya022.botcommands.internal.commands.ratelimit.declaration
 
 import io.github.freya022.botcommands.api.commands.builder.RateLimitBuilder
-import io.github.freya022.botcommands.api.commands.ratelimit.RateLimitContainer
 import io.github.freya022.botcommands.api.commands.ratelimit.RateLimitInfo
 import io.github.freya022.botcommands.api.commands.ratelimit.RateLimiterFactory
 import io.github.freya022.botcommands.api.commands.ratelimit.bucket.BucketFactory
 import io.github.freya022.botcommands.api.commands.ratelimit.declaration.RateLimitManager
 import io.github.freya022.botcommands.api.core.BContext
 import io.github.freya022.botcommands.api.core.service.getService
+import io.github.freya022.botcommands.internal.commands.ratelimit.RateLimitContainer
 
 internal class RateLimitManagerImpl internal constructor(override val context: BContext) : RateLimitManager() {
     private val container = context.getService<RateLimitContainer>()
